Worker email fetching and releasing
Implements queue-like semantics in place of an actual queue
implementation. Workers will fetch an email that's "available" based on
a couple set criteria, attempt to send it, and then release it back to
the pool. In the event that it's complete, no further processing will
take place. Otherwise, the email will be picked up based on its
"incomplete" status as well as it not being "claimed" by any worker.

Currently, only the fetch and release works. Additionally, the worker
spins up multiple goroutines to prove the concurrency works.
Unfortunately, signals are no longer being handled correctly, so it's
not possible to kill the process without kill -9.

To be continued...