Cleanup + README + ops work
Sets up the worker to be built and run in a container along side the
other components in the Compose environment. Additionally, starts to
flesh out the README with instructions for running the project. Lastly,
makes the interactio between the API and CLI make more sense, with the
API returning JSON blobs and the CLI doing the work of prettifying with
prettytable, as well as removing a couple extraneous methods I was using
for testing.