Devproject (#1)
* Directory structure and brainstorming

* Boilerplate Setup + Flask App

Implements a basic flask app, mongo insert for email documents, and a
dirt-simple CLI for POST'ing emails to our endpoint. Additionally adds
Dockerfile and docker-compose functionality for testing and dev.

* Fleshing out the API

Adds
* Support for CC and BCC
* Routes for fetching status
* Output formatting / views

* Starting on the go worker

* Worker email fetching and releasing

Implements queue-like semantics in place of an actual queue
implementation. Workers will fetch an email that's "available" based on
a couple set criteria, attempt to send it, and then release it back to
the pool. In the event that it's complete, no further processing will
take place. Otherwise, the email will be picked up based on its
"incomplete" status as well as it not being "claimed" by any worker.

Currently, only the fetch and release works. Additionally, the worker
spins up multiple goroutines to prove the concurrency works.
Unfortunately, signals are no longer being handled correctly, so it's
not possible to kill the process without kill -9.

To be continued...

* Fixes + SMTP

Implements basic SMTP and sends emails from the worker process based on
documents in the database. Additionally, sets up channels for each of
the child processes so that they properly quit when the parent receives
a SIGTERM.

* Starts to add email formatting

* SMTP Robustness

Adds handling for various error states from the remote server, as well
as correct status fields and some handling for failures. Lastly,
corrects some incorrect formatting and ensures that To, CC and BCC are
handled correctly.

* Implements proper selection intervals

The email sending logic now correctly uses the updated_at timestamp to
retry emails on an N second interval dependent on the failure condition.

* Adds some new CLI commands

Implements commands to fetch emails and view their status from the
Python API.

* Adds random server selection

Servers are now a database model and are queried for when emails are
sent. This allows any number of servers to be defined as endpoints for
the workers. Additionally, does some code cleanup and adds robustness

* Test scaffolding and lint cleanups

* Cleanup + README + ops work

Sets up the worker to be built and run in a container along side the
other components in the Compose environment. Additionally, starts to
flesh out the README with instructions for running the project. Lastly,
makes the interactio between the API and CLI make more sense, with the
API returning JSON blobs and the CLI doing the work of prettifying with
prettytable, as well as removing a couple extraneous methods I was using
for testing.

* Unit tests + refactoring

Adds unit tests for unit-testable things in app.py, as well as
refactoring the CLI into a CLI and API client to make it easier to test
and also leverage in functional tests. Lastly, adds tests for the
refactored client code

* Functional Tests

Adds several functional tests written in Python to exercise all the
positive paths around sending, fetching and deleting emails. Also adds a
couple functional paths to exercise deterministic failures.

* More Functional Tests

Adds functional tests in Python using the mailgun client provided in tbe
project

* Golang tests and final touches

Adds a suite of golang unit tests for the worker code, and does some
minor cleanup and refactoring of the code base.

* Addressing feedback

Reworks the quit channel code to simply rely on a close() rather than
sending a unique channel and bool to each. Also fixes the worker
container and the go src import to help reconcile paths in IDEs

* Fixes install_python task

The install_python task erroneously omitted the test-requirements.txt
from the pip install command.